Added a bunch of new manifests!
I am currently on the road to learning puppet on Windows, and have found
that there is very little help for noobs along the way. My hope is these
manifests are very simple and to the point, so more people can learn
Puppet on Windows, and eventually there will be many available resources
for this in the future.